+++ src/Helpers/DotNotationHelper.php
@@ -6,13 +6,13 @@ namespace Blumilk\BLT\Helpers;
 
 class DotNotationHelper
 {
-    public static function getValueUsingDotNotation(array $array, string $path)
+    public static function getValueUsingDotNotation(array $array, string $path): mixed
     {
         $keys = explode(".", $path);
 
         foreach ($keys as $key) {
             if (!isset($array[$key])) {
-                return;
+                return null;
             }
 
             $array = $array[$key];
